==> notebooks/single-country/2_generate_grids.ipynb <==
---
title: Single Country Rollout Part 1 (Generate Roll-out Grids) 
format:
  html:
    code-fold: false
---

%matplotlib inline
%reload_ext autoreload
%autoreload 2

import sys

sys.path.append("../../../")
import os
from pathlib import Path

import geopandas as gpd
import pandas as pd

from povertymapping import settings
from povertymapping.iso3 import get_region_name
from povertymapping.rollout_grids import get_region_filtered_bingtile_grids

# Generate Roll-out Grids 

This notebook generates the rollout grid tiles over the country. The output file is used as an input for Step 3, where we will run the trained model over the set of grids.

The generated grids are set at 2.4km (zoom level 14), matching the grids used during model training. The grids are also filtered to only include populated areas based on Meta's High Resolution Settlement Layer (HRSL) data.

## Set country-specific parameters

COUNTRY_CODE = "tl"
ROLLOUT_DATE = None

DATA_DIR = settings.DATA_DIR.resolve()
ROLLOUT_DIR = DATA_DIR/f"rollout/{COUNTRY_CODE}"
ROLLOUT_DIR.mkdir(parents=True,exist_ok=True)

REGION = get_region_name(COUNTRY_CODE, code="alpha-2").lower()
ADMIN_LVL = "ADM2"
ZOOM_LEVEL = 14

## Generate Grids

#papermill_description="Generate rollout grids"
admin_grids_gdf = get_region_filtered_bingtile_grids(
    REGION, admin_lvl=ADMIN_LVL, quadkey_lvl=ZOOM_LEVEL, use_cache=True
)

## Explore Populated Grids

admin_grids_gdf.info()

admin_grids_gdf.head(2)

# Uncomment to view the generated grid on an interactive map
# admin_grids_gdf.explore()

## Save to file

#papermill_description="Save rollout grids"
ROLLOUT_DATE = "-".join(os.getcwd().split("/")[-2].split("-")[:3]) if ROLLOUT_DATE is None else ROLLOUT_DATE
grid_save_path = f"{ROLLOUT_DIR}/{ROLLOUT_DATE}-{COUNTRY_CODE}-rollout-grids.geojson"
admin_grids_gdf.to_file(grid_save_path, driver="GeoJSON", index=False)
